refactor(tools): extract BoundedOutputReader from ShellTool (task group 1)
Move BoundedDrainAsync + the tail-ring helpers out of ShellTool into a
reusable BoundedOutputReader (Netclaw.Actors/Tools), so background_job and
file_read can share the same bounded, allocation-flat capture instead of
copy-pasting the ring math (the #1293 review's altitude finding).

- DrainToWindowAsync(reader, budget, ct): the head+tail window, unchanged
  behavior; now takes a CancellationToken (ShellTool passes None to preserve
  the pipe-drain semantics; file readers can pass a real token).
- DrainCaptureAsync(reader, captureMax, inlineBudget, ct): retains a capture
  body for the spill plus a smaller inline window from one bounded pass.
- Window(string, budget): pure head+tail of an in-memory string.

The reader is a pure leaf — no redaction, no file IO — so callers own those.
ShellTool and the benchmark now call BoundedOutputReader; the drain tests move
to BoundedOutputReaderTests with added Window/DrainCapture coverage. No
behavior change to shell_execute. 40 reader + shell tests pass; slopwatch clean.
