+++ src/Netclaw.Actors/Tools/ShellTool.cs
@@ -3,7 +3,6 @@
 //      Copyright (C) 2026 - 2026 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
-using System.Buffers;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Text;
@@ -159,11 +158,11 @@ public sealed partial class ShellTool : NetclawTool<ShellTool.Params>
             // pipe write-ends open, so a blocked pipe read cannot be interrupted
             // by a token; killing the process is what closes the pipes.
             //
-            // BoundedDrainAsync reads into a head+tail window bounded by
+            // BoundedOutputReader reads into a head+tail window bounded by
             // MaxOutputChars but continues draining after the cap is reached so
             // the pipe never fills up and deadlocks a still-running child.
-            var stdoutTask = BoundedDrainAsync(process.StandardOutput, _config.MaxOutputChars);
-            var stderrTask = BoundedDrainAsync(process.StandardError, _config.MaxOutputChars);
+            var stdoutTask = BoundedOutputReader.DrainToWindowAsync(process.StandardOutput, _config.MaxOutputChars, CancellationToken.None);
+            var stderrTask = BoundedOutputReader.DrainToWindowAsync(process.StandardError, _config.MaxOutputChars, CancellationToken.None);
 
             try
             {
@@ -190,7 +189,7 @@ public sealed partial class ShellTool : NetclawTool<ShellTool.Params>
                 {
                     // If the OS refuses the kill, the child may keep stdout/stderr
                     // open forever. Close our read ends so cancellation still
-                    // returns promptly instead of hanging in BoundedDrainAsync.
+                    // returns promptly instead of hanging in the pipe drain.
                     killClosedPipes = false;
                     Debug.WriteLine($"shell_execute: process kill skipped — {ex.Message}");
                     process.StandardOutput.Dispose();
@@ -238,151 +237,9 @@ public sealed partial class ShellTool : NetclawTool<ShellTool.Params>
         }
     }
 
-    /// <summary>
-    /// Drains <paramref name="reader"/> into a head+tail window bounded by
-    /// <paramref name="maxChars"/>. Chars beyond the cap are discarded but the
-    /// pipe continues to be read so a still-running child never deadlocks on a
-    /// full pipe buffer. Returns the captured text and whether it was truncated.
-    /// </summary>
-    /// <remarks>
-    /// Allocation is O(<paramref name="maxChars"/>), not O(total output): the
-    /// scratch read buffer is pooled, the tail ring is allocated only if the head
-    /// fills, and reads go through the <see cref="ValueTask{T}"/> overload so a
-    /// pipe that already has data buffered completes synchronously without a
-    /// per-chunk <see cref="Task"/> allocation. A child that prints hundreds of MB
-    /// is drained with a handful of KB of managed allocations — the property that
-    /// keeps a memory-limited daemon from being OOM-killed (see #1293).
-    /// </remarks>
-    internal static async Task<(string Text, bool Truncated)> BoundedDrainAsync(
-        TextReader reader, int maxChars)
-    {
-        if (maxChars <= 0)
-        {
-            // Explicit opt-out: a non-positive cap disables truncation and reads
-            // the whole stream. The config schema enforces MaxOutputChars >= 1, so
-            // production never reaches this — it exists only for programmatic
-            // callers that deliberately pass 0/negative to capture full output.
-            var all = await reader.ReadToEndAsync(CancellationToken.None);
-            return (all, false);
-        }
-
-        // Split the budget: first half for the head, second half for the tail.
-        // Odd maxChars gives the extra char to the head. Computed without (maxChars
-        // + 1) so a near-int.MaxValue cap can't overflow to a negative headCap.
-        var headCap = maxChars / 2 + maxChars % 2;
-        var tailCap = maxChars / 2;
-
-        var head = new StringBuilder(Math.Min(headCap, 4096));
-
-        // Tail ring buffer, allocated lazily on first overflow past the head. The
-        // common case — output under the cap — never fills the head, so it never
-        // pays for the tail window at all.
-        char[]? tailBuf = null;
-        var tailStart = 0;  // index of the oldest retained char in the ring
-        var tailLen = 0;    // chars currently retained (<= tailCap)
-
-        long totalChars = 0; // total chars seen across all reads; long so a multi-GB
-                             // flood can't overflow the truncation check to a false negative
-
-        // Transient scratch buffer for the read loop: pooled so a long drain
-        // doesn't allocate it per call and it never lands on the LOH.
-        var buf = ArrayPool<char>.Shared.Rent(4096);
-        try
-        {
-            int read;
-            // ReadAsync(Memory<char>) returns a non-allocating ValueTask when the
-            // read completes synchronously (data already buffered) — unlike the
-            // Task<int> char[] overload, which allocates once per chunk.
-            while ((read = await reader.ReadAsync(buf.AsMemory(), CancellationToken.None)) > 0)
-            {
-                totalChars += read;
-                var span = buf.AsSpan(0, read);
-
-                if (head.Length < headCap)
-                {
-                    var headChunk = Math.Min(headCap - head.Length, span.Length);
-                    head.Append(span[..headChunk]);
-                    span = span[headChunk..];
-                }
-
-                if (span.IsEmpty || tailCap == 0)
-                    continue;
-
-                tailBuf ??= new char[tailCap];
-                AppendToTailRing(tailBuf, span, ref tailStart, ref tailLen);
-            }
-        }
-        finally
-        {
-            // clearArray: the scratch buffer held raw stdout/stderr (possibly
-            // secrets); wipe it before returning to the shared pool.
-            ArrayPool<char>.Shared.Return(buf, clearArray: true);
-        }
-
-        // Truncation only when total chars exceeded the full budget (head+tail),
-        // meaning some middle chars were discarded.
-        var truncated = totalChars > maxChars;
-
-        // Reconstruct in place on `head`: when truncated, the discarded middle is
-        // marked with a separator; otherwise head + tail is the full output. Reusing
-        // `head` avoids a second StringBuilder and re-copying the head.
-        if (truncated)
-            head.Append("\n...\n");
-        if (tailBuf is not null && tailLen > 0)
-            AppendRing(head, tailBuf, tailStart, tailLen);
-        return (head.ToString(), truncated);
-    }
-
-    /// <summary>
-    /// Writes <paramref name="span"/> into a ring buffer that retains only the
-    /// most recent <c>ring.Length</c> chars. Uses block copies (at most two per
-    /// call) rather than a per-char loop, so draining a very chatty child stays
-    /// cheap regardless of how much it prints.
-    /// </summary>
-    private static void AppendToTailRing(char[] ring, ReadOnlySpan<char> span, ref int start, ref int len)
-    {
-        var cap = ring.Length;
-
-        if (span.Length >= cap)
-        {
-            // This span alone fills (or overfills) the window: only its last `cap`
-            // chars can survive. One contiguous copy, ring reset.
-            span[^cap..].CopyTo(ring);
-            start = 0;
-            len = cap;
-            return;
-        }
-
-        var writePos = (start + len) % cap;
-        var first = Math.Min(span.Length, cap - writePos);
-        span[..first].CopyTo(ring.AsSpan(writePos));
-        if (first < span.Length)
-            span[first..].CopyTo(ring); // remainder wraps to the front
-
-        var newLen = len + span.Length;
-        if (newLen > cap)
-        {
-            // Overwrote the oldest chars: advance start past them.
-            start = (start + (newLen - cap)) % cap;
-            len = cap;
-        }
-        else
-        {
-            len = newLen;
-        }
-    }
-
-    /// <summary>Appends a ring buffer's retained chars, oldest-first, to <paramref name="sb"/>.</summary>
-    private static void AppendRing(StringBuilder sb, char[] ring, int start, int len)
-    {
-        var first = Math.Min(len, ring.Length - start);
-        sb.Append(ring, start, first);
-        if (first < len)
-            sb.Append(ring, 0, len - first);
-    }
-
-    // Retained for compatibility with tests that call it directly; the main
-    // execution path no longer uses this — BoundedDrainAsync caps at read time.
+    // Retained for compatibility with tests/benchmark that call it directly; the
+    // main execution path no longer uses this — output is bounded at read time by
+    // BoundedOutputReader.
     internal static string TruncateOutput(string output, int maxChars)
     {
         if (output.Length <= maxChars)
